refactor: update dependencies and restructure database handling
- Updated dependencies in package.json, including Clerk SDK and Drizzle ORM.
- Introduced a new db/index.ts file for database initialization and schema export.
- Removed obsolete migrate.ts file and related migration logic.
- Enhanced schema definitions in schema.ts with additional constraints and types.
- Cleaned up index.ts by removing unused middleware and streamlining server setup.
- Deleted unused GitHub authentication middleware and related token attachment logic.
- Refactored socket authentication middleware to utilize the new db structure.
- Updated project service to use the new database initialization method.
- Modified Clerk utility functions to align with the new Clerk SDK structure.
- Adjusted type imports to reflect changes in the Clerk SDK.
