--- backend/server/drizzle.config.ts
@@ -1,12 +1,0 @@
-import "dotenv/config"
-import type { Config } from "drizzle-kit"
-import { defineConfig } from "drizzle-kit"
-
-export default defineConfig({
-  out: "./drizzle",
-  schema: "./src/db/schema.ts",
-  dialect: "postgresql",
-  dbCredentials: {
-    url: process.env.DATABASE_URL!,
-  },
-}) satisfies Config
