--- backend/server/drizzle/0001_unusual_archangel.sql
@@ -1,1 +1,0 @@
-ALTER TABLE "sandbox" ADD COLUMN "repositoryId" text;
