--- backend/server/drizzle/meta/_journal.json
@@ -1,20 +1,0 @@
-{
-  "version": "7",
-  "dialect": "postgresql",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "7",
-      "when": 1739612282205,
-      "tag": "0000_superb_nomad",
-      "breakpoints": true
-    },
-    {
-      "idx": 1,
-      "version": "7",
-      "when": 1739612320600,
-      "tag": "0001_unusual_archangel",
-      "breakpoints": true
-    }
-  ]
-}
