+++ backend/server/package.json
@@ -6,47 +6,40 @@
   "scripts": {
     "build": "npx tsc",
     "start": "node dist/src/index.js",
-    "dev": "nodemon src/index.ts",
-    "generate": "npx drizzle-kit generate",
-    "migrate": "tsx src/migrate.ts"
+    "dev": "nodemon src/index.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@clerk/clerk-sdk-node": "^4.13.23",
-    "@octokit/core": "^6.1.4",
+    "@clerk/backend": "^1.34.0",
     "@paralleldrive/cuid2": "^2.2.2",
     "concurrently": "^8.2.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "drizzle-orm": "^0.39.2",
+    "drizzle-orm": "^0.44.1",
     "e2b": "^1.2.5-autopause.1",
     "express": "^4.19.2",
-    "itty-router-extras": "^0.4.6",
-    "jiti": "^2.4.2",
-    "jzip": "^1.0.0",
     "pg": "^8.13.1",
     "rate-limiter-flexible": "^5.0.3",
     "simple-git": "^3.25.0",
     "socket.io": "^4.7.5",
     "ssh2": "^1.15.0",
     "tar": "^7.4.3",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "zod-openapi": "^4.2.4"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
-    "@types/itty-router-extras": "^0.4.3",
     "@types/jszip": "^3.4.1",
     "@types/node": "^20.12.7",
     "@types/pg": "^8.11.11",
     "@types/socket.io": "^3.0.1",
     "@types/ssh2": "^1.15.0",
     "@types/tar": "^6.1.13",
-    "drizzle-kit": "^0.30.4",
+    "drizzle-kit": "^0.31.1",
     "nodemon": "^3.1.0",
     "ts-node": "^10.9.2",
-    "tsx": "^4.19.2",
     "typescript": "^5.4.5"
   }
 }
