+++ backend/server/src/index.ts
@@ -5,8 +5,6 @@ import fs from "fs"
 import { createServer } from "http"
 import { Server, Socket } from "socket.io"
 
-import { attachAuthToken } from "./middleware/attachAuthToken"
-import { requireAuth } from "./middleware/clerkAuth"
 import { socketAuth } from "./middleware/socketAuth"
 import { ConnectionManager } from "./services/ConnectionManager"
 import { DokkuClient } from "./services/DokkuClient"
@@ -43,9 +41,6 @@ dotenv.config()
 const app: Express = express()
 const port = process.env.PORT || 4000
 app.use(cors())
-
-// Apply Clerk authentication middleware and attach auth token middleware to all API routes
-app.use("/api", requireAuth, attachAuthToken)
 
 const httpServer = createServer(app)
 const io = new Server(httpServer, {
@@ -161,9 +156,6 @@ io.on("connection", async (socket) => {
     handleErrors("Error connecting:", e, socket)
   }
 })
-
-// REST API routes:
-app.use(express.json())
 
 // Start the server
 httpServer.listen(port, () => {
